Moved panels into separate subfolder
