+++ src/ui/panels/skill-descriptions.js
